Restructure of common functions
