+++ webviz_4d/_datainput/common.py
@@ -1,4 +1,4 @@
-""" Collection of common functions """
+""" Collection of some functions used by the 4D viewer or data preparation scripts """
 
 import os
 import io
@@ -45,6 +45,29 @@ def get_config_item(config,key):
                 pass             
     
     return value
+    
+    
+def get_map_defaults(configuration, n):
+    map_defaults = []
+
+    settings_file = configuration["settings"]
+    settings_file = get_full_path(settings_file)
+    
+    settings = read_config(settings_file)
+    interval = settings["map_settings"]["default_interval"]
+
+    if not "-" in interval[0:8]:
+        date1 = interval[0:4] + "-" + interval[4:6] + "-" + interval[6:8]
+        date2 = interval[9:13] + "-" + interval[13:15] + "-" + interval[15:17]
+        interval = date1 + "-" + date2
+
+    for i in range(0, n):
+        key = "map" + str(i + 1) + "_defaults"
+        defaults = settings["map_settings"][key]
+        defaults["interval"] = interval
+        map_defaults.append(defaults)
+
+    return map_defaults    
     
 
 def decode_filename(file_path, delimiter):
@@ -113,6 +136,15 @@ def _find_number(surfacepath, txt):
         number = filename[i : i + j]
 
     return number
+    
+    
+def check_number(string):
+    
+    for i in range(len(string)):
+        if string[i] not in "0123456789":
+            return False
+    
+    return True     
     
     
 def get_dates(interval):
@@ -220,37 +252,42 @@ def get_well_colors(settings):
     """ Return well colors from a configuration """   
 
     return settings["well_colors"]
+    
+    
+def get_update_dates(wellfolder):
+    update_dates = {}
+    
+    try:
+        well_date_file = os.path.join(wellfolder,".welldata_update.yaml")
 
+        with open(well_date_file, "r") as stream:
+            well_meta_data = yaml.safe_load(stream)
+            
+        well_update = well_meta_data[0]["welldata"]["update_time"]
+        update_dates["well_update_date"] = well_update.strftime("%Y-%m-%d %H:%M:%S")
+    except:
+        update_dates["well_update_date"] = ''
+    
+    try:
+        prod_date_file = os.path.join(wellfolder,".production_update.yaml")
 
-def extract_wellbore_metadata(directory):
-    """ Return all well metadata found in yaml files in a given folder """
-    well_info = []
-    interval_info = []
-
-    yaml_files = glob.glob(str(directory) + "/.*yaml", recursive=True)
-
-    for yaml_file in yaml_files:
-        with open(yaml_file, "r") as stream:
-            data = yaml.safe_load(stream)
-            # print('data',data)
-
-            well_info.append(data[0])
-
-            if len(data) > 1 and data[1]:
-                for item in data[1:]:
-                    interval_info.append(item)
-
-    well_info_df = json_normalize(well_info)
-
-    interval_df = json_normalize(interval_info)
-
-    if not interval_df.empty:
-        interval_df.sort_values(
-            by=["interval.wellbore", "interval.mdTop"], inplace=True
-        )
-
-    return well_info_df, interval_df
-
+        with open(prod_date_file, "r") as stream:
+            production_meta_data = yaml.safe_load(stream)
+          
+        #print(production_meta_data)
+        first_date = production_meta_data[0]["production"]["start_date"].strftime("%Y-%m-%d")
+        last_date = production_meta_data[0]["production"]["last_date"].strftime("%Y-%m-%d")
+        
+        update_dates["production_first_date"] = first_date           
+        update_dates["production_last_date"] = last_date   
+    except:  
+        update_dates["production_first_date"] = ''           
+        update_dates["production_last_date"] = ''  
+        
+    #print("Update dates", update_dates)        
+    
+    return update_dates    
+    
 
 def get_well_metadata(metadata_df, wellbore_name, item):
     """ Return selected metadata for a selected wellbore """
@@ -422,6 +459,54 @@ def get_full_path(item):
         full_path = None    
         
     return full_path 
+    
+    
+def get_plot_label(configuration, interval):
+    difference_mode = "normal"
+    labels = []
+
+    dates = [
+        interval[:4] + interval[5:7] + interval[8:10],
+        interval[11:15] + interval[16:18] + interval[19:21],
+    ]
+
+    for date in dates:
+        # date = convert_date(date)
+        try:
+            labels_dict = configuration["date_labels"]
+            label = labels_dict[int(date)]
+        except:
+            label = date[:4] + "-" + date[4:6] + "-" + date[6:8]
+
+        labels.append(label)
+
+    if difference_mode == "normal":
+        label = labels[0] + " - " + labels[1]
+    else:
+        label = labels[1] + " - " + labels[0]
+
+    return label    
+    
+    
+def get_colormap(configuration, attribute):
+    colormap = None
+    minval = None
+    maxval = None
+
+    try:
+        attribute_dict = configuration[attribute]
+        # print("attribute_dict", attribute_dict)
+        colormap = attribute_dict["colormap"]
+        minval = attribute_dict["min_value"]
+        minval = attribute_dict["max_value"]
+    except:
+        try:
+            map_settings = configuration("map_settings")
+            colormap = map_settings("default_colormap")
+        except:
+            print("No default colormaps found for ", attribute)
+
+    return colormap, minval, maxval    
     
     
 def main():
